Local Verification added
Added script for local verification, which includes methods for checking the distance between two points on the map and returning whether we consider the pin local to the user who created it.
When setting new pins, the app now uses the checkLocal() function to set a Boolean on the pin's local value.